+++ src/drone_controller/calculation_layer/thrust_calculator.py
